[bin] Add helpers to copy a GitLab shortcode derived from an URL
About 10 times a day I want to copy a GitLab reference (e.g.
group/project#issue) to my clipboard. This helper makes it easier to
derive the shortcode from an GitLab URL and it works for merge_requests
as well as issues or epics.

The second helper will transform the shortcode into a Markdown link, so
it can easily pasted into emails or Mattermost.