[nix] Fuck it. Fish is nice, but bash is what everyone else uses
I feel this commit message could be a blog post on it's own. For now
let's leave it with: fish shell is amazing, but I need to jump back and
forth between it's syntax and standard POSIX bash scripts that I have to
write, debug and exchange with co-workers on a daily basis.

Even simple for loops aren't portable, so fish isn't helping but adding
mental workload for everyone involved. It's gotten so much better
through the years though – and almost everything comes _included_.

For a brief moment I thought about going back to `zsh`, but it's just
not as snappy as bash or fish, so… bash it is.

This change isn't complete yet, but it ports my aliases and completions
back to simple bash files I can source. More to follow.