[emacs] Rework org-mode Personio reporting to reflect state in org-tree
Up until now it worked but I needed to manually keep track of the days I
already reported to Personio. This change will add a org-property
`PERSONIO_REPORTED` to the respective days tree, and
`(bascht/ts-for-report-table)` will format the table accordingly, so
that only unreported days get a elisp: link to report new times.

This was a nice journey into `(async-start)` and how the scope binding
works between the async lambda and the resulting callback.